Move audio to an audio package and add support for web
